Working with DropDown Menu.

+++ GuitarShop.BLL/Dtos/AddToCartDto.cs
@@ -3,6 +3,6 @@ namespace GuitarShop.BLL.Dtos;
 
 public class AddToCartDto
 {
-    public long UserId { get; set; } 
-    public long ProductId { get; set; } 
+    public int UserId { get; set; } 
+    public int ProductId { get; set; } 
 }
