+++ GuitarShop.DAL/Entities/Cart.cs
@@ -4,8 +4,8 @@ namespace GuitarShop.DAL.Entities;
 
 public class Cart
 {
-    public long Id { get; set; }
-    public long UserId { get; set; }
+    public int Id { get; set; }
+    public int UserId { get; set; }
     
     public virtual ICollection<CartItem> CartItems { get; set; }
