+++ GuitarShop/Models/CartItemViewModel.cs
@@ -2,8 +2,8 @@
 
 public class CartItemViewModel
 {
-    public long Id { get; set; }
-    public long CartId { get; set; }
+    public int Id { get; set; }
+    public int CartId { get; set; }
     public ProductViewModel Product { get; set; }
     public int Quantity { get; set; }
     public decimal TotalPrice { get; set; }
